Use of GetTickCount64()

+++ ut/ut.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <string>
+#include <Windows.h>
 
-const int SECONDS_PER_SECOND = 1;
-const int SECONDS_PER_MINUTE = 60;
-const int SECONDS_PER_HOUR = 3600;
-const int SECONDS_PER_DAY = SECONDS_PER_HOUR * 24;
+const ULONGLONG SECONDS_PER_SECOND = 1;
+const ULONGLONG SECONDS_PER_MINUTE = 60;
+const ULONGLONG SECONDS_PER_HOUR = 3600;
+const ULONGLONG SECONDS_PER_DAY = SECONDS_PER_HOUR * 24;
 
-std::string FormatSecondsLong(unsigned int sec);
-std::string FormatSecondsShort(unsigned int sec);
+std::string FormatSecondsLong(ULONGLONG sec);
+std::string FormatSecondsShort(ULONGLONG sec);
